+++ tbtb-oa/src/main/java/com/tbtaobao/cloud/service/PersonService.java
@@ -11,6 +11,6 @@ public interface PersonService {
 
     public int create(Person person);
 
-    public Person getPersonById(@Param("id") Long id);
+    public Person getPersonById(@Param("id") Integer id);
 
 }
